Refactor: Introduction of TestCase base class.
Split the responsability of "Defining a test" to the one that
actually "run the test". Each test is unique in its definition,
but still follow a pattern, and running a test is common to every
test, because all of them must be executed. Therefore, the
commonnalities (the base pattern/structure and the runner) are
abstracted into a TestCase Base class, that each specific test must
override.