Add support for a log string.
Implementation of teardown() feature with yet another flag yields an
apparent limitation with those flags, that they miss the crutial aspect
of keeping track of the order of execution of the test stages: Setup(),
Running() and teardown(). Therefore, introducing a log string is pretty
much appropriate as it, unlike the flags, properly captures execution
order.
Finally, there is no more need of the flags, so they can be removed with
any inherent functionnality.