Counting/Collecting Failed results.
When does a test execution fail? Mostly in 2 cases: Either there is an
execution error during setup() or run(), either the method under test
runs properly but the assertion associated with it is not satisfied.
In both cases, an exception is thrown, but i feel like i have to deal
with these 2 separately.
Fpr now, let focus on when the assertion is not correct. In this case,
we catch the assertion, consider this as test failure and increment the
failures count. As the assertion is caught, we can no longer rely on it,
instead, we need to directly print/report the test results on the console.