Report collected results: Refactoring & polishing.
TestCase.run() started to get complex, so we extracted some critical
parts of it into its own method.
We refactored TestResults to give it a more suitable interface for
collaborations.
Finally, asserting on the reports of the tests executions, doesn't make
anymore sense as these are mostly assertions on the test runner's
behavior, not on the behavior of the methods under test. So we replace them
with proper assertions on the behavior of the methods under test, just
as any normal test case will do.