Catch exceptions in setUp().
Exceptions in setUp() are now caught and the appropriate behavior (at
least the one implemented by python's unittest) is now correctly
manifested, as no further test is ran after any exception in setUp().
The exception is caught and the error is reported in the test execution
summary. Even tearDown is not executed because of the partial state in
which the blown up in setUp() leaves us (see "destructor hell").
Side note: I've introduced an "unpythonic", anti "PEP8" and anti "Zen
of python" ugly more than 80 lines long ternary if-else condition. I've
accepted to commit this sin as I know this is a temporary situation that
will be fixed further while implementing test suites features.